image/requirements.txt
clip==0.2.0
einops==0.8.1
huggingface_hub==0.27.1
numpy==2.2.6
omegaconf==2.3.0
opencv_python==4.12.0.88
Pillow==9.5.0
pytorch_lightning==2.5.0.post0
Requests==2.32.4
scikit_learn==1.7.0
scipy==1.16.0
torch_fidelity==0.3.0
torchmetrics==1.6.1
torchvision==0.20.1+cu121
tqdm==4.66.5
wandb
webdataset==0.2.100

packaging==21.3
matplotlib
seaborn